+++ src/routes/usersRoute.ts
@@ -1,5 +1,5 @@
 import express from "express"
-import userController from "../controllers/usersController"
+import userController from "../controllers/userController"
 
 const router = express.Router()
 
@@ -7,7 +7,7 @@ router.post('/duplication', userController.userAvailableCheck)
 router.post('/signup', userController.createUser)
 router.post('/signin', userController.signInUser)
 router.post('/kakaoLogin', userController.kakaoLogin)
-router.post('/kakaoLogout', userController.kakaoLogout) // 굳이 카카오와 일반 로그아웃을 구별?
+router.post('/kakaoLogout', userController.kakaoLogout)
 
 
 export default router
